Fix array-bounds warning on GCC 9
Generated C++ constructors for messages optimize setting multiple
contiguous primitive fields using memset.  Currently, that call to
memset is undefined behavior, because it writes beyond the bounds of
the first field.  GCC 9 detects this and issues a warning.

This commit fixes the warning by rephrasing the memset in terms of
offsets within the message class, copying the strategy from a Linux
kernel fix for the same problem [1].

[1] https://lore.kernel.org/lkml/20190517092502.GA22779@gmail.com/